Refactor 'Role' entity to new_entity
