+++ source code/backend/src/main/java/Backend/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package Backend.repository;
 
-import Backend.entity.Role;
+import Backend.new_entity.Role;
 import Backend.utilities.ERole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
