+++ source code/backend/src/main/java/Backend/service/UserServiceImpl.java
@@ -3,7 +3,7 @@ package Backend.service;
 import Backend.config.JwtUtils;
 import Backend.dto.UserExport;
 import Backend.entity.PasswordResetToken;
-import Backend.entity.Role;
+import Backend.new_entity.Role;
 import Backend.entity.User;
 import Backend.repository.PasswordResetTokenRepository;
 import Backend.repository.UserRepository;
